[#4336] Remove html5 shim from stats extension.
html5.js was removed in commit #3705 but still referenced in
the stats extension. Currently this causes an error when the page
loads in CKAN 2.8 installs.

Removing this will remove support for older IE versions which
will only cause some issues for CKAN 2.8 users with the old
bootstrap 2 theme being used.